Abstracting deprecation test generation
