+++ .dev/tests/jest/helpers.js
@@ -1,4 +1,8 @@
-import { getBlockTransforms } from '@wordpress/blocks';
+/**
+ * External dependencies
+ */
+import { omit } from 'lodash';
+import { registerBlockType, unregisterBlockType, createBlock, getBlockTransforms, serialize, parse } from '@wordpress/blocks';
 
 /**
  * A simplified version of the prefix trigger located in the RichText component.
@@ -20,3 +24,72 @@ export const performPrefixTransformation = ( blockName, prefix, content ) => {
 
 	return block;
 };
+
+/**
+ * Generate tests for each defined deprecation of a block.
+ *
+ * @param {String} blockName The registered block name.
+ * @param {Object} blockSettings The registered block settings.
+ * @param {Object} blockVariations The used attributes and value varitions.
+ */
+export const testDeprecatedBlockVariations = ( blockName, blockSettings, blockVariations ) =>
+	blockSettings.deprecated.map( ( deprecated, index ) => {
+		describe( `${ blockName } deprecation ${ index }`, () => {
+			// Make variables accessible for all tests.
+			let deprecatedBlock;
+
+			beforeEach( () => {
+				unregisterBlockType( blockName );
+
+				// Register the deprecated block.
+				const deprecatedSettings = Object.assign(
+					{}, omit( blockSettings, [ 'attributes', 'save', 'deprecated' ] ),
+					{
+						attributes: deprecated.attributes,
+						save: deprecated.save,
+					}
+				);
+				registerBlockType( blockName, { category: 'common', ...deprecatedSettings } );
+
+				// Create the block with the minimum attributes.
+				deprecatedBlock = createBlock( blockName );
+			} );
+
+			it( 'should deprecate old version', () => {
+				const deprecatedSerialized = serialize( deprecatedBlock );
+
+				// Unregister the deprecated block version.
+				unregisterBlockType( blockName );
+
+				// Register the current block version.
+				registerBlockType( blockName, { category: 'common', ...blockSettings } );
+
+				const blocks = parse( deprecatedSerialized );
+
+				expect(
+					blocks.every( block => block.isValid )
+				).toBe( true );
+			} );
+
+			Object.keys( deprecated.attributes ).map( ( attribute ) => {
+				blockVariations[ attribute ].map( variation => {
+					it( `should support attribute.${ attribute } set to '${ variation }'`, () => {
+						deprecatedBlock.attributes[ attribute ] = variation;
+						const deprecatedSerialized = serialize( deprecatedBlock );
+
+						// Unregister the deprecated block version.
+						unregisterBlockType( blockName );
+
+						// Register the current block version.
+						registerBlockType( blockName, { category: 'common', ...blockSettings } );
+
+						const blocks = parse( deprecatedSerialized );
+
+						expect(
+							blocks.every( block => block.isValid )
+						).toBe( true );
+					} );
+				} );
+			} );
+		} );
+	} );
